refactor: change locations background to be the same of the rest of the app
