+++ src/pages/dashboard/Locations.jsx
@@ -1,9 +1,12 @@
 import MapLocations from "../../components/dashboard/MapLocations";
+import styles from "./Dashboard.module.css";
 
 function Locations() {
   return (
-    <div>
-      <MapLocations />
+    <div className={styles.backgroundPages}>
+      <div className={`container mt-5 ${styles.agendaPanel}`}>
+        <MapLocations />
+      </div>
     </div>
   );
 }
